backend/requirements.txt
django
djangorestframework
django-cors-headers
pillow
django-filter
